Measure brightness in lux
This is to get rid of the <LivisiSensor> is using native unit of measurement '%' which is not a valid unit for the device class ('illuminance') warning. Unfortunately Luminance is provided in percent by Livisi, and there is no real info on how to convert the values, so i measured a few data points and found info online on a cap at 400lx, so we can approximate the value, which should be good enough for automations that depend on brightness.